access control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,10 @@ use DB;
 
 class PostsController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -77,6 +81,9 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         return view('posts.edit')->with('post',$post);
     }
 
@@ -111,6 +118,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized Page');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Removed');
     }
